assignment/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic-ai>=0.0.25
pydantic-settings>=2.2.0
openai>=1.0.0
httpx>=0.27.0
python-multipart>=0.0.9
pypdf>=4.0.0
python-docx>=1.1.0